top level cli cleanups
