+++ cmd/nabu/test.go
@@ -7,27 +7,19 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/internetofwater/nabu/internal/config"
 	"github.com/internetofwater/nabu/internal/synchronizer"
 
 	log "github.com/sirupsen/logrus"
 )
 
-type TestCmd struct{}
-
-func Test(cfgStruct config.NabuConfig) error {
-	client, err := synchronizer.NewSynchronizerClientFromConfig(cfgStruct)
-	if err != nil {
-		return err
-	}
-
-	exists, err := client.S3Client.Client.BucketExists(context.Background(), cfgStruct.Minio.Bucket)
+func Test(ctx context.Context, client *synchronizer.SynchronizerClient) error {
+	exists, err := client.S3Client.Client.BucketExists(ctx, client.S3Client.DefaultBucket)
 	if err != nil {
 		return err
 	}
 
 	if !exists {
-		return fmt.Errorf("default bucket %s does not exist", cfgStruct.Minio.Bucket)
+		return fmt.Errorf("default bucket %s does not exist", client.S3Client.DefaultBucket)
 	}
 
 	log.Info("s3 test passed")
